refactor: move fetching logic to hook file
